Refactor read parameters for AgentProvider
Before this change the parameters were split in patr for setting up the
woodpecker agent and a second part for provider specific parameters.
But as the AWS provider does not need parameters for the agent at all
(the agent must have been setup outside of picus before), all parameters
for an AgentProvider are now read as part of the AgentProvider.